Home Dividido em Componentes
Home dividido em componentes para melhorar legibilidade e manuntenção do código
