+++ frontend/src/pages/Home/Home.css
@@ -5,302 +5,7 @@
   color: #1e293b;
 }
 
-/* Header */
-.home-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px 32px;
-  background-color: #ffffff;
-  border-bottom: 1px solid #e2e8f0;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
-}
 
-
-.auth-buttons {
-  display: flex;
-  gap: 12px;
-}
-
-.user-profile-nav {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.user-greeting {
-  font-size: 14px;
-  color: #475569;
-}
-
-.role-badge {
-  padding: 4px 10px;
-  border-radius: 12px;
-  font-size: 11px;
-  font-weight: 700;
-  text-transform: uppercase;
-}
-
-.role-badge.prestador,
-.role-badge.worker,
-.role-badge.admin {
-  background-color: #fef3c7;
-  color: #92400e;
-}
-
-.role-badge.user {
-  background-color: #e0e7ff;
-  color: #3730a3;
-}
-
-.role-badge.prestador {
-  background-color: #dcfce7;
-  color: #166534;
-}
-
-.btn-logout {
-/* Botões padrão */
-.btn-primary {
-  padding: 8px 16px;
-  background-color: #2563eb;
-  color: #ffffff;
-  border-radius: 6px;
-  text-decoration: none;
-  font-weight: 600;
-  font-size: 14px;
-}
-
-.btn-secondary {
-  padding: 8px 16px;
-  color: #2563eb;
-  border: 1px solid #2563eb;
-  border-radius: 6px;
-  text-decoration: none;
-  font-weight: 600;
-  font-size: 14px;
-}
-
-.btn-logout {
-  padding: 6px 14px;
-  background-color: #ef4444;
-  color: #ffffff;
-  border: none;
-  border-radius: 6px;
-  font-weight: 600;
-  cursor: pointer;
-}
-
-/* Hero Section */
-.hero-section {
-  background: linear-gradient(135deg, #1e293b 0%, #0f172a 100%);
-  color: #ffffff;
-  padding: 64px 20px;
-  text-align: center;
-}
-
-.hero-content {
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-.hero-content h1 {
-  font-size: 36px;
-  font-weight: 800;
-  margin-bottom: 16px;
-  line-height: 1.2;
-}
-
-.hero-subtitle {
-  font-size: 18px;
-  color: #94a3b8;
-  margin-bottom: 32px;
-}
-
-.hero-ctas {
-  display: flex;
-  justify-content: center;
-  gap: 16px;
-  flex-wrap: wrap;
-}
-
-.welcome-card {
-  width: 100%;
-  max-width: 600px;
-  background: #ffffff;
-  padding: 32px;
-  border-radius: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
-.btn-hero-primary,
-.btn-hero-worker {
-  padding: 14px 28px;
-  font-size: 16px;
-  font-weight: 600;
-  border-radius: 8px;
-  border: none;
-  cursor: pointer;
-  transition: transform 0.1s, background-color 0.2s;
-}
-
-.btn-hero-primary {
-  background-color: #2563eb;
-  color: #ffffff;
-}
-
-.btn-hero-primary:hover {
-  background-color: #1d4ed8;
-}
-
-.status-box {
-  margin-top: 24px;
-  padding: 16px 20px;
-  background-color: #f9fafb;
-  border: 1px solid #e5e7eb;
-  border-radius: 8px;
-.btn-hero-worker {
-  background-color: #10b981;
-  color: #ffffff;
-}
-
-.btn-hero-worker:hover {
-  background-color: #059669;
-}
-
-.btn-hero-secondary {
-  padding: 14px 28px;
-  font-size: 16px;
-  font-weight: 600;
-  border-radius: 8px;
-  color: #ffffff;
-  border: 1px solid #475569;
-  text-decoration: none;
-}
-
-/* Categories Section */
-.categories-section {
-  max-width: 1000px;
-  margin: 48px auto;
-  padding: 0 20px;
-  text-align: center;
-}
-
-.categories-section h2 {
-  font-size: 24px;
-  color: #0f172a;
-  margin-bottom: 8px;
-}
-
-.status-box code {
-  background-color: #e5e7eb;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-family: monospace;
-  color: #1f2937;
-}
-
-.provider-action {
-  margin-top: 24px;
-  padding: 20px;
-  border: 1px dashed #0f766e;
-  border-radius: 16px;
-  background: #ecfdf5;
-}
-
-.provider-action h3 {
-  margin-top: 0;
-  margin-bottom: 8px;
-  color: #065f46;
-}
-
-.provider-action p {
-  margin: 0 0 16px;
-  color: #047857;
-  font-size: 0.95rem;
-}
-
-.provider-status-card {
-  margin-top: 24px;
-  padding: 20px;
-  border: 1px solid #bbf7d0;
-  border-radius: 16px;
-  background: #f0fdf4;
-}
-
-.provider-status-badge {
-  display: flex;
-  align-items: flex-start;
-  gap: 14px;
-}
-
-.provider-status-badge h3 {
-  margin-top: 0;
-  margin-bottom: 4px;
-  color: #166534;
-  font-size: 1.1rem;
-}
-
-.provider-status-badge p {
-  margin: 0;
-  color: #15803d;
-  font-size: 0.95rem;
-}
-
-.badge-icon {
-  font-size: 1.5rem;
-  line-height: 1;
-}
-
-.btn-primary {
-  padding: 12px 18px;
-  background-color: #0f766e;
-  color: #ffffff;
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  font-weight: 700;
-  transition: background-color 0.2s;
-}
-
-.btn-primary:hover {
-  background-color: #115e59;
-.categories-section p {
-  color: #64748b;
-  margin-bottom: 32px;
-}
-
-.categories-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
-  gap: 20px;
-}
-
-.category-card {
-  background-color: #ffffff;
-  border: 1px solid #e2e8f0;
-  border-radius: 12px;
-  padding: 24px 16px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-  cursor: pointer;
-  transition: transform 0.2s, box-shadow 0.2s;
-}
-
-.category-card:hover {
-  transform: translateY(-4px);
-  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
-  border-color: #3b82f6;
-}
-
-.category-icon {
-  font-size: 32px;
-}
-
-.category-name {
-  font-weight: 600;
-  font-size: 14px;
-  color: #334155;
-}
 
 /* Responsividade */
 @media (max-width: 640px) {
