+++ frontend/src/pages/Home/components/HomeCategories/HomeCategories.css
@@ -1,0 +1,100 @@
+/* Categories Section */
+.categories-section {
+  max-width: 1000px;
+  margin: 48px auto;
+  padding: 0 20px;
+  text-align: center;
+}
+
+.categories-section h2 {
+  font-size: 24px;
+  color: #0f172a;
+  margin-bottom: 8px;
+}
+
+.status-box code {
+  background-color: #e5e7eb;
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-family: monospace;
+  color: #1f2937;
+}
+
+.provider-action {
+  margin-top: 24px;
+  padding: 20px;
+  border: 1px dashed #0f766e;
+  border-radius: 16px;
+  background: #ecfdf5;
+}
+
+.provider-action h3 {
+  margin-top: 0;
+  margin-bottom: 8px;
+  color: #065f46;
+}
+
+.provider-action p {
+  margin: 0 0 16px;
+  color: #047857;
+  font-size: 0.95rem;
+}
+
+.badge-icon {
+  font-size: 1.5rem;
+  line-height: 1;
+}
+
+.btn-become-provider {
+  padding: 12px 18px;
+  background-color: #0f766e;
+  color: #ffffff;
+  border: none;
+  border-radius: 10px;
+  cursor: pointer;
+  font-weight: 700;
+  transition: background-color 0.2s;
+}
+
+.btn-become-provider:hover {
+  background-color: #115e59;
+}
+.categories-section p {
+  color: #64748b;
+  margin-bottom: 32px;
+}
+
+.categories-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
+  gap: 20px;
+}
+
+.category-card {
+  background-color: #ffffff;
+  border: 1px solid #e2e8f0;
+  border-radius: 12px;
+  padding: 24px 16px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+  cursor: pointer;
+  transition: transform 0.2s, box-shadow 0.2s;
+}
+
+.category-card:hover {
+  transform: translateY(-4px);
+  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
+  border-color: #3b82f6;
+}
+
+.category-icon {
+  font-size: 32px;
+}
+
+.category-name {
+  font-weight: 600;
+  font-size: 14px;
+  color: #334155;
+}
